+++ SAMD51_FINAL/My Project/Apps/Tasks/ModemTask/include/ModemResponseHandles.h
@@ -13,7 +13,7 @@
 #include <stdbool.h>
 #include "Apps/Tasks/ModemTask/include/ModemAtCommandSet.h"
 
-void modemResponseHandler(AT_CMD_TYPE cmd,uint8_t* response, uint8_t length);
+void modemResponseHandler(AT_CMD_TYPE cmd,uint8_t* response, uint16_t length);
 void mdmResp_AtRespHandler(uint8_t* response, uint8_t length);
 void mdmResp_IMEIRespHandler(uint8_t* response, uint8_t length);
 void mdmResp_CarrierRespHandler(uint8_t* response, uint8_t length);
